code/requirements.txt
structlog==22.1.0
python-dateutil==2.8.2